move code to the appropriate location
