+++ src/sentry/templates/sentry/emails/transactions.html
@@ -32,7 +32,7 @@
         <tr>
             <th>{% blocktrans %} Repeating Spans ({{ num_repeating_spans }}) {% endblocktrans %}</th>
             <td>
-                <code>{{ repeating_spans|truncatechars:50 }}</code>
+                <code>{{ repeating_spans|truncatechars:75 }}</code>
             </td>
         </tr>
         {% endif %}
