+++ src/sentry/web/frontend/debug/debug_performance_issue.py
@@ -2,15 +2,13 @@ import pytz
 from django.utils.safestring import mark_safe
 from django.views.generic import View
 
-from sentry.api.serializers.models.event import get_entries, get_problems
 from sentry.event_manager import EventManager
-from sentry.models import GroupHash, Organization, Project, Rule
-from sentry.notifications.utils import get_group_settings_link, get_rules
+from sentry.models import Organization, Project, Rule
+from sentry.notifications.utils import get_group_settings_link, get_rules, get_transaction_data
 from sentry.testutils.helpers import override_options
 from sentry.testutils.helpers.datetime import before_now
 from sentry.utils import json
 from sentry.utils.samples import load_data
-from sentry.web.helpers import render_to_string
 
 from .mail import MailPreview
 
@@ -49,50 +47,6 @@ COMMIT_EXAMPLE = """[
 
 
 class DebugPerformanceIssueEmailView(View):
-    def perf_to_email_html(self, spans, problem):
-
-        # I'm aware this is not great duplication
-        def get_span_evidence_value_problem(problem):
-            value = "no value"
-            if not problem.op and problem.desc:
-                value = problem.desc
-            if problem.op and not problem.desc:
-                value = problem.op
-            if problem.op and problem.desc:
-                value = f"{problem.op} - {problem.desc}"
-            return value
-
-        def get_span_evidence_value(span):
-            value = "no value"
-            if not span.get("op") and span.get("description"):
-                value = span.get("description")
-            if span.get("op") and not span.get("description"):
-                value = span.get("op")
-            if span.get("op") and span.get("description"):
-                op = span.get("op")
-                desc = span.get("description")
-                value = f"{op} - {desc}"
-            return value
-
-        parent_span = None
-        repeating_spans = None
-        for span in spans:
-            if problem.parent_span_ids[0] == span.get("span_id"):
-                parent_span = span
-            if problem.offender_span_ids[0] == span.get("span_id"):
-                repeating_spans = span
-            if parent_span is not None and repeating_spans is not None:
-                break
-
-        context = {
-            "transaction_name": get_span_evidence_value_problem(problem),
-            "parent_span": get_span_evidence_value(parent_span),
-            # "preceding_span": "SELECT idk FROM idk_what_a_preceding_span_is",
-            "repeating_spans": get_span_evidence_value(repeating_spans),
-            "num_repeating_spans": str(len(problem.offender_span_ids)),
-        }
-        return render_to_string("sentry/emails/transactions.html", context)
-
     @override_options({"performance.issues.all.problem-creation": 1.0})
     @override_options({"performance.issues.all.problem-detection": 1.0})
     @override_options({"performance.issues.n_plus_one_db.problem-creation": 1.0})
@@ -118,17 +72,7 @@ class DebugPerformanceIssueEmailView(View):
 
         rule = Rule(id=1, label="Example performance rule")
 
-        entries = get_entries(perf_event, None)
-        spans = []
-        if len(entries):
-            spans = [entry.get("data") for entry in entries[0] if entry.get("type") == "spans"][0]
-
-        matched_problem = None
-        for problem in get_problems([perf_event]):
-            if problem.problem.fingerprint == GroupHash.objects.get(group=perf_group).hash:
-                matched_problem = problem
-
-        spans = self.perf_to_email_html(spans, matched_problem.problem)
+        transaction_data = get_transaction_data(perf_event)
 
         # XXX: this interface_list code needs to be the same as in
         # get_interface_list in src/sentry/notifications/utils/__init__.py
@@ -158,6 +102,6 @@ class DebugPerformanceIssueEmailView(View):
                 "tags": perf_event.tags,
                 "project_label": project.slug,
                 "commits": json.loads(COMMIT_EXAMPLE),
-                "transaction_data": [("Span Evidence", mark_safe(spans), None)],
+                "transaction_data": [("Span Evidence", mark_safe(transaction_data), None)],
             },
         ).render(request)
